requirements.txt
attrs==17.4.0
bleach==2.1.2
decorator==4.2.1
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.8
django-autocomplete-light==3.2.10
django-debug-toolbar==1.9.1
django-extensions==1.9.9
django-filter==1.1.0
django-formtools==2.1
django-logentry-admin==1.0.3
django-material==1.2.2
django-reversion==2.0.13
django-shortuuidfield==0.1.3
django-test-without-migrations==0.6
django-viewflow==1.2.0
entrypoints==0.2.3
gunicorn==19.4.1
html5lib==1.0.1
ipdb==0.11
ipykernel==4.8.2
ipython
ipython-genutils==0.2.0
ipywidgets==7.1.2
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
MarkupSafe==1.0
mistune==0.8.3
model-mommy==1.5.1
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.0
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
pluggy==0.6.0
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
py==1.5.2
Pygments==1.6
pytest==3.4.0
pytest-django==3.1.2
python-dateutil==2.6.1
python-decouple==3.1
pytz==2017.3
pyzmq==17.0.0
qtconsole==4.3.1
raven==6.5.0
Send2Trash==1.5.0
shortuuid==0.5.0
simplegeneric==0.8.1
six==1.11.0
# sqlformatter==1.3
sqlparse==0.1.11
static3==0.7.0
terminado==0.8.1
testpath==0.3.1
toml==0.9.4
tornado==4.5.3
traitlets==4.3.2
typing==3.6.2
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==3.3.1
widgetsnbextension==3.1.4
xlwt==1.3.0
